Modified calculation of SNR DIETRICH1 (Neff)
The effective number of coils estimated from the noise images is a real
number, which does not make sense in Dietrich's estimation of the
correction factor for noise distribution.

Neff is replaced by round(Neff).

This is not really important since this method is not considered the best
way to estimate SNR...