Auto QA processing
Script in beta version, running on vivelyne's desktop, with safety steps
in place to avoid deleting files before checking that the results are
properly generated and data are archived.

The defaults file can be easily modified so the automated QA script can run
from elsewhere (final version: on the fmriserver2).

The compression step (for archiving), currently achieved using tar.gz
compression with Matlab, may be modified in the final version.